Made a bit nicer code for the sample
