+++ app/src/main/java/com/lb/apkparser/AbstractZipInputStreamFilter.kt
@@ -1,0 +1,6 @@
+package com.lb.apkparser
+
+abstract class AbstractZipInputStreamFilter {
+    abstract fun getNextEntryName(): String?
+    abstract fun getBytesFromCurrentEntry(): ByteArray?
+}
