+++ app/src/main/java/com/lb/apkparser/MainActivity.kt
@@ -6,17 +6,11 @@ import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
-import net.dongliu.apk.parser.parser.*
-import net.dongliu.apk.parser.struct.AndroidConstants
-import net.dongliu.apk.parser.struct.resource.ResourceTable
 import java.io.FileInputStream
-import java.nio.ByteBuffer
 import java.util.*
 import java.util.zip.ZipInputStream
 import kotlin.concurrent.thread
 
-@Suppress("DEPRECATION")
-fun PackageInfo.versionCodeCompat(): Long = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) longVersionCode else versionCode.toLong()
 private const val VALIDATE_RESOURCES = true
 
 class MainActivity : AppCompatActivity() {
@@ -46,8 +40,8 @@ class MainActivity : AppCompatActivity() {
 //                        val manifestXml = it.manifestXml
 //                        Log.d("AppLog", "")
 //                    }
-                    ZipInputStream(FileInputStream(apkFilePath)).use {
-                        val apkInfo = getApkInfo(locale, it, true, VALIDATE_RESOURCES)
+                    ZipInputStreamFilter(ZipInputStream(FileInputStream(apkFilePath))).use {
+                        val apkInfo = ApkInfo.getApkInfo(locale, it, true, VALIDATE_RESOURCES)
                         when {
                             apkInfo == null -> Log.e("AppLog", "can't parse apk:$apkFilePath")
                             apkInfo.apkType == null -> Log.e("AppLog", "can\'t get apk type: $apkFilePath ")
@@ -63,7 +57,7 @@ class MainActivity : AppCompatActivity() {
                                     when {
                                         packageInfo.packageName != apkMeta.packageName -> Log.e("AppLog", "apk package name is different for $apkFilePath : correct one is: ${packageInfo.packageName} vs found: ${apkMeta.packageName}")
                                         packageInfo.versionName != apkMeta.versionName -> Log.e("AppLog", "apk version name is different for $apkFilePath : correct one is: ${packageInfo.versionName} vs found: ${apkMeta.versionName}")
-                                        packageInfo.versionCodeCompat() != apkMeta.versionCode -> Log.e("AppLog", "apk version code is different for $apkFilePath : correct one is: ${packageInfo.versionCodeCompat()} vs found: ${apkMeta.versionCode}")
+                                        versionCodeCompat(packageInfo) != apkMeta.versionCode -> Log.e("AppLog", "apk version code is different for $apkFilePath : correct one is: ${versionCodeCompat(packageInfo)} vs found: ${apkMeta.versionCode}")
                                         label != labelOfLibrary -> Log.e("AppLog", "apk label is different for $apkFilePath : correct one is: $label vs found: $labelOfLibrary")
                                         else -> {
                                             Log.d("AppLog", "apk data of $apkFilePath : ${apkMeta.packageName}, ${apkMeta.versionCode}, ${apkMeta.versionName}, $labelOfLibrary, ${apkMeta.icon}, ${apkMetaTranslator.iconPaths}")
@@ -78,8 +72,8 @@ class MainActivity : AppCompatActivity() {
                 }
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
                     packageInfo.applicationInfo.splitPublicSourceDirs?.forEach { apkFilePath ->
-                        ZipInputStream(FileInputStream(apkFilePath)).use {
-                            val apkInfo = getApkInfo(locale, it, true, false)
+                        ZipInputStreamFilter(ZipInputStream(FileInputStream(apkFilePath))).use {
+                            val apkInfo = ApkInfo.getApkInfo(locale, it, true, false)
                             when {
                                 apkInfo == null -> Log.e("AppLog", "can\'t parse apk:$apkFilePath")
                                 apkInfo.apkType == null -> Log.e("AppLog", "can\'t get apk type: $apkFilePath")
@@ -91,7 +85,7 @@ class MainActivity : AppCompatActivity() {
                                     val apkMetaTranslator = apkInfo.apkMetaTranslator
                                     when {
                                         packageInfo.packageName != apkMeta.packageName -> Log.e("AppLog", "apk package name is different for $apkFilePath : correct one is: ${packageInfo.packageName} vs found: ${apkMeta.packageName}")
-                                        packageInfo.versionCodeCompat() != apkMeta.versionCode -> Log.e("AppLog", "apk version code is different for $apkFilePath : correct one is: ${packageInfo.versionCodeCompat()} vs found: ${apkMeta.versionCode}")
+                                        versionCodeCompat(packageInfo) != apkMeta.versionCode -> Log.e("AppLog", "apk version code is different for $apkFilePath : correct one is: ${versionCodeCompat(packageInfo)} vs found: ${apkMeta.versionCode}")
                                         else -> Log.d("AppLog", "apk data of $apkFilePath : ${apkMeta.packageName}, ${apkMeta.versionCode}, ${apkMeta.versionName}, ${apkMeta.name}, ${apkMeta.icon}, ${apkMetaTranslator.iconPaths}")
                                     }
 
@@ -109,100 +103,7 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun getApkInfo(locale: Locale, zipInputStream: ZipInputStream, requestParseManifestXmlTagForApkType: Boolean = false, requestParseResources: Boolean = false): ApkInfo? {
-        var manifestBytes: ByteArray? = null
-        var resourcesBytes: ByteArray? = null
-        while (true) {
-            val zipEntry = zipInputStream.nextEntry ?: break
-            when (zipEntry.name) {
-                AndroidConstants.MANIFEST_FILE -> {
-                    manifestBytes = zipInputStream.readBytes()
-                    if (!requestParseResources || resourcesBytes != null)
-                        break
-                }
-                AndroidConstants.RESOURCE_FILE -> {
-                    if (!requestParseResources)
-                        continue
-                    resourcesBytes = zipInputStream.readBytes()
-                    if (manifestBytes != null)
-                        break
-                }
-            }
-        }
-        if (manifestBytes == null) {
-//            Log.e("AppLog", "could not find manifest file for $apkFilePath")
-            return null
-        }
-        val xmlTranslator = XmlTranslator()
-        val resourceTable: ResourceTable =
-                if (resourcesBytes == null)
-                    ResourceTable()
-                else {
-                    val resourceTableParser = ResourceTableParser(ByteBuffer.wrap(resourcesBytes))
-                    resourceTableParser.parse()
-                    resourceTableParser.resourceTable
-                    //                this.locales = resourceTableParser.locales
-                }
-        val apkMetaTranslator = ApkMetaTranslator(resourceTable, locale)
-        val binaryXmlParser = BinaryXmlParser(ByteBuffer.wrap(manifestBytes), resourceTable)
-        binaryXmlParser.locale = locale
-        binaryXmlParser.xmlStreamer = CompositeXmlStreamer(xmlTranslator, apkMetaTranslator)
-        binaryXmlParser.parse()
-        var apkType: ApkInfo.ApkType? = null
-        if (requestParseManifestXmlTagForApkType) {
-            val apkMeta = apkMetaTranslator.apkMeta
-            val isSplitApk = !apkMeta.split.isNullOrEmpty()
-            if (isSplitApk)
-                apkType = ApkInfo.ApkType.SPLIT
-            else {
-                //standalone or base of split apks
-                val isDefinitelyBaseApkOfSplit = apkMeta.isSplitRequired
-                if (isDefinitelyBaseApkOfSplit)
-                    apkType = ApkInfo.ApkType.BASE_OF_SPLIT
-                else {
-                    val manifestXml = xmlTranslator.xml
-                    apkType = ApkInfo.ApkType.STANDALONE
-                    try {
-                        XmlTag.getXmlFromString(manifestXml)?.innerTagsAndContent?.forEach { manifestXmlItem: Any ->
-                            if (manifestXmlItem is XmlTag && manifestXmlItem.tagName == "application") {
-                                val innerTagsAndContent = manifestXmlItem.innerTagsAndContent
-                                        ?: return@forEach
-                                for (applicationXmlItem: Any in innerTagsAndContent) {
-                                    if (applicationXmlItem is XmlTag && applicationXmlItem.tagName == "meta-data"
-                                            && applicationXmlItem.tagAttributes?.get("name") == "com.android.vending.splits") {
-                                        apkType = ApkInfo.ApkType.BASE_OF_SPLIT_OR_STANDALONE
-                                    }
-                                    if (applicationXmlItem is XmlTag && applicationXmlItem.tagName == "meta-data"
-                                            && applicationXmlItem.tagAttributes?.get("name") == "instantapps.clients.allowed" &&
-                                            applicationXmlItem.tagAttributes!!["value"] != "false") {
-                                        apkType = ApkInfo.ApkType.BASE_OF_SPLIT_OR_STANDALONE
-                                    }
-                                    if (applicationXmlItem is XmlTag && applicationXmlItem.tagName == "meta-data"
-                                            && applicationXmlItem.tagAttributes?.get("name") == "com.android.vending.splits.required") {
-                                        val isSplitRequired = applicationXmlItem.tagAttributes!!["value"] != "false"
-                                        if(!isSplitRequired)
-                                        Log.e("AppLog", "!isSplitRequired")
-//                                        apkType = if (isSplitRequired) ApkInfo.ApkType.BASE_OF_SPLIT else ApkInfo.ApkType.BASE_OF_SPLIT_OR_STANDALONE
-                                        apkType = ApkInfo.ApkType.BASE_OF_SPLIT
-                                        break
-                                    }
-                                }
-                            }
-                            return@forEach
-                        }
-                    } catch (e: Exception) {
-                        e.printStackTrace()
-                        Log.e("AppLog", "failed to get apk type: $e")
-                    }
-                }
-            }
-        }
-        return ApkInfo(xmlTranslator, apkMetaTranslator, apkType)
-    }
-
-    class ApkInfo(val xmlTranslator: XmlTranslator, val apkMetaTranslator: ApkMetaTranslator, val apkType: ApkType?) {
-        enum class ApkType {
-            STANDALONE, BASE_OF_SPLIT, SPLIT, BASE_OF_SPLIT_OR_STANDALONE
-        }
+    companion object {
+        fun versionCodeCompat(packageInfo: PackageInfo) = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) packageInfo.longVersionCode else packageInfo.versionCode.toLong()
     }
 }
