+++ build.gradle
@@ -9,11 +9,7 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:4.1.0-alpha05"
 //        classpath 'com.android.tools.build:gradle:3.5.3'
-
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
     }
 }
