+++ lib/build.gradle
@@ -9,8 +9,8 @@ java {
 
 dependencies {
     implementation 'org.jetbrains:annotations:18.0.0'
-    // https://mvnrepository.com/artifact/org.bouncycastle/bcprov-jdk15on
 //    implementation 'com.google.code.findbugs:jsr305:3.0.2'
+    // https://mvnrepository.com/artifact/org.bouncycastle/bcprov-jdk15on
     implementation 'org.bouncycastle:bcprov-jdk15on:1.65'
     implementation 'org.bouncycastle:bcpkix-jdk15on:1.65'
     implementation 'junit:junit:4.12'
